requirements.txt
django
djangorestframework
djangorestframework-simplejwt 
pytest
pytest-django
pytest-cov 
django-cors-headers
